backend/requirements.txt
aniso8601==9.0.1
backcall==0.2.0
backports.entry-points-selectable==1.1.1
black==21.11b0
cfgv==3.3.1
click==8.0.3
decorator==5.1.0
distlib==0.3.3
dynaconf==3.1.7
filelock==3.4.0
flake8==4.0.1
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
gunicorn==20.1.0
identify==2.4.0
ipdb==0.13.9
ipython==7.30.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.21.4
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
pre-commit==2.15.0
prompt-toolkit==3.0.23
ptyprocess==0.7.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.10.0
PySDL2==0.9.9
pyserial==3.5
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
scipy==1.7.3
six==1.16.0
toml==0.10.2
tomli==1.2.2
traitlets==5.1.1
typing_extensions==4.0.0
virtualenv==20.10.0
wcwidth==0.2.5
Werkzeug==2.0.2
